refactor: isolate language related logic to hook
